feedback-updates: added express routers for endpoints, added 5 recent additions-updates to landing page upon load
