+++ public/index.html
@@ -15,7 +15,7 @@
       <form action="#" class="js-search-form" role="search">
         <fieldset class="band-list-fieldset">
           <legend>Do I Have That Album Already?</legend>
-          <label for="band-list">Choose a band from the list to see if you have a specific album from them already.</label>
+          <label for="band-list" class="band-list-label">Choose a band from the list to see if you have a specific album from them already.</label>
           <select id="band-list" class="band-list" aria-label="band selection"></select><br />
           <input type="button" class="add-new-album-toggle" value="Add A New Album">                 
         </fieldset>
@@ -35,6 +35,7 @@
         </fieldset>
       </form>    
       <section class="js-search-results" aria-live="assertive">
+        <section class="js-most-recent"></section>
       </section>
     </main>        
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.1.0/jquery.min.js"></script>
